Create Day-10-Make-Array-Zero-by-Subtracting-Equal-Amounts
The minimumOperations function takes a vector nums as input, which represents a sequence of integers.
The function creates a set s using the elements from nums. This set will automatically remove any duplicates and sort the elements in ascending order.
If the smallest element in the set is 0, it means we cannot perform any operation on it because it cannot be reduced further. Therefore, we subtract 1 from the size of the set before returning it.
If the smallest element is not 0, it means all elements are unique, and we can return the size of the set as it is.
The returned value represents the minimum number of operations required to make all elements in nums unique.