Create Day-11-Maximum-Number-of-Coins-You-Can-Get
The maxCoins function takes a vector piles as input, representing the number of coins in each pile.
The function first sorts the piles vector in ascending order using the sort function from the <algorithm> library.
It then calculates the number of operations (k) needed to select the largest one-third of the piles. Since each operation involves selecting the pile with the maximum number of coins, we need to consider only the largest one-third of the piles.
The function initializes a variable result to store the total number of coins obtained.
It then iterates over the piles vector starting from the k-th pile and incrementing by 2 in each iteration. This is done to select every other pile in the largest one-third.
In each iteration, it adds the number of coins in the current pile to the result variable.
Finally, the function returns the total number of coins obtained (result).